requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnslib==0.9.26
dnspython==2.8.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
gyp==0.1
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.11
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.14.0
keras==3.12.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==8.10.0
namex==0.1.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
openai==2.32.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.16.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4.1
python-dotenv==1.2.1
PyYAML==5.4.1
requests==2.32.5
rich==14.2.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
systemd-python==234
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.5.0
wadllib==1.3.6
Werkzeug==3.1.3
wrapt==2.0.1
zipp==1.0.0
